dupe-cascade: refactor: add fix loop duplicates
